requirements.txt
python-telegram-bot
spotipy
beautifulsoup4
